Unit tests (Part 1)
- New: Unit tests for model::players
- New: Unit tests for model::clubs

- Changed: Now all `name_color` fields are u64

- Fix: serde-rename fields to camelCase on ClubMember (thanks, unit tests!)
- Fix: serde-rename 'result' to 'battle' on Battle (thanks, unit tests!)
- Fix: Now `deserialize_number_from_string` works on hexadecimal nums ("0x......")

+++ src/model/rankings/players.rs
@@ -281,7 +281,7 @@ pub struct PlayerRanking {
 
     /// The player's name color. Defaults to `0xffffff` (white).
     #[serde(default = "oxffffff_default")]
-    pub name_color: usize,
+    pub name_color: u64,
 }
 
 /// Represents the club in a player's ranking (a [`PlayerRanking`] object). Since the only data
